pyproject.toml
[tool.poetry]
name = "ctf-django"
version = "0.1.0"
description = ""
authors = ["David Tillemans <david.tillemans@cryptable.org>"]
license = "GNU General Public License v3 (GPL-3)"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
cryptography = "^42.0.3"
pycryptodome = "^3.20.0"
django-bootstrap5 = "^23.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
